Create missing permissions attribute variable for admin_dashboard
The permissions attribute for the admin_dashboard is not set
if the POLICY_CHECK_FUNCITON is set. This was a bug fix backported into
Mitaka (https://review.openstack.org/#/c/123741/). So this caused
horizon-customization to break since the empty list being referenced
at index 0 raises an error ( permission = []; permission[0] is bad.)
This fixes the problem by setting admin_dashboard.permissions