Create domain work around for federated users
I filed a bug here: https://bugs.launchpad.net/horizon/+bug/1627062

Federated admins and users have trouble using horizon because
they are in the federated domain. They are not able to list/create projects,
groups, and users because horizon filters by domain. This code
will see if the user is federated and then set the domain to
the current scoped project's domain.